requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
django-filter
Pillow
